refactor: separate discover and query source module
