+++ enterprise/frontend/src/embedding-sdk-package/data-app-query-sync/canonical.ts
