+++ enterprise/frontend/src/embedding-sdk-package/data-app-query-sync/tests/setup.ts
