+++ enterprise/frontend/src/embedding-sdk-package/data-app-query-sync/types.ts
